Revert "utolsó simítások"
This reverts commit 965ee187a3ea50937af0727ac12bac9653e51ea3, reversing
changes made to 4e04efcab4bdafe6f1134c7a23be96ccca0f7524.

+++ autok/SearchFilters.cs
@@ -16,8 +16,8 @@ namespace autok
             set { selectedMarka = value; OnPropertyChanged(nameof(SelectedMarka)); }
         }
 
-        private int selectedEvjarat;
-        public int SelectedEvjarat
+        private double selectedEvjarat;
+        public double SelectedEvjarat
         {
             get => selectedEvjarat;
             set { selectedEvjarat = value; OnPropertyChanged(nameof(SelectedEvjarat)); }
